refactor of init.sqf
* init.sqf is now postInit
* some have been relocated to core, which has been refactored
* adds two new settings isServer and isPublic for future usage to core
* removed some unnecessary network trafic
