+++ @twc_framework/addons/twc_core/CfgEventHandlers.hpp
@@ -1,7 +1,7 @@
 class Extended_PostInit_EventHandlers {
-	class twc_core_eh {
-		serverInit = "[] call TWC_Core_fnc_serverEventHandlers;";
-		clientInit = "[] call TWC_Core_fnc_clientEventHandlers;";
+	class TWC_Core {
+		serverInit = "[] call TWC_Core_fnc_serverPostInit;";
+		clientInit = "[] call TWC_Core_fnc_clientPostInit;";
 	};
 };
 
@@ -10,5 +10,11 @@ class Extended_InitPost_EventHandlers {
 		class TWC_Core_NightGear_EH {
 			init = "_this call TWC_Core_fnc_initNightGear;";
 		};
+	};
+};
+
+class Extended_PreInit_EventHandlers {
+	class TWC_Core {
+		init = "_this call TWC_Core_fnc_clientPreInit;";
 	};
 };
