--- @twc_framework/addons/twc_core/functions/fn_clientEventHandlers.sqf
@@ -1,5 +1,0 @@
-params ["_player"];
-
-if (!local _player) exitWith {};
-
-["TWC_AddedToGroup", { _this call TWC_Core_fnc_addedToGroup; }] call CBA_fnc_addEventHandler;
