+++ @twc_framework/addons/twc_framework/CfgEventHandlers.hpp
@@ -1,0 +1,27 @@
+class Extended_InitPost_EventHandlers {
+	class CAManBase {
+		class twc_moduleInitPos_eh {
+			init = "[(_this select 0)] execVM 'twc_framework\init.sqf';";
+		};
+		
+		class twc_functionInit {
+			init = "_this call twc_fnc_init;";
+		};
+		
+		class twc_AIItemCheckInit {
+			init = "_this call twc_fnc_confiscateBadItems;";
+		};
+	};
+	
+	class TWC_Module_WaveSpawn {
+		class TWC_ModuleWaveSpawnInit {
+			init = "_this call TWC_fnc_moduleWaveSpawnInit;";
+		};
+	};
+};
+
+class Extended_PostInit_EventHandlers {
+	class TWC_Framework {
+		init = "_this call TWC_fnc_postInit;";
+	};
+};
