+++ @twc_framework/addons/twc_framework/cfgFunctions.hpp
@@ -63,6 +63,9 @@ class CfgFunctions {
 			/** AMBIENT AIRCRAFT **/
 			class ambientFlyBy {};
 			class ambientFormationFlyBy {};
+			
+			/** Init **/
+			class postInit {};
 		};
 		
 		class modules {
